Stats: Move summary-chart
