+++ client/my-sites/stats/stats-summary-chart/index.jsx
@@ -11,7 +11,7 @@ var React = require( 'react' ),
  */
 var observe = require( 'lib/mixins/data-observe' ),
 	ElementChart = require( 'components/chart' ),
-	StatTab = require( './stats-tabs/tab' ),
+	StatTab = require( '../stats-tabs/tab' ),
 	analytics = require( 'analytics' ),
 	Card = require( 'components/card' );
