+++ client/my-sites/stats/summary.jsx
@@ -13,7 +13,7 @@ var observe = require( 'lib/mixins/data-observe' ),
 	StatsModule = require( './stats-module' ),
 	StatsStrings = require( './stats-strings' )(),
 	Countries = require( './stats-countries' ),
-	SummaryChart = require( './module-summary-chart' ),
+	SummaryChart = require( './stats-summary-chart' ),
 	VideoPlayDetails = require( './stats-video-details' );
 
 module.exports = React.createClass( {
